Fix compiler error

+++ misc.h
@@ -65,7 +65,7 @@ static inline const char *Timestamp2String(int64_t ts, uint8_t divisor)
 	static char buf[3][16];
 	static int idx = 0;
 
-	if (ts == AV_NOPTS_VALUE) {
+	if (ts == (int64_t) AV_NOPTS_VALUE) {
 		return "--:--:--.---";
 	}
